# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vinyl Radar API",
    "description": "국내 바이닐 발매·재고 정보 통합 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/push/subscribe": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Subscribe",
        "description": "구독을 등록한다.\n\n같은 엔드포인트로 다시 구독하면 **오류가 아니라 갱신**이다. 브라우저는 권한을\n다시 물을 때마다 같은 구독을 돌려주므로, 중복 요청이 정상 경로다.",
        "operationId": "subscribe_v1_push_subscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "push"
        ],
        "summary": "Unsubscribe",
        "description": "구독을 해지한다.\n\n행을 지우지 않고 **비활성화**한다 — 이미 보낸 기록(`notification_deliveries`)이\n참조하고 있고, 무엇을 언제 보냈는지는 남아야 한다.\n없는 엔드포인트를 해지해도 204 다 (멱등).",
        "operationId": "unsubscribe_v1_push_subscribe_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushUnsubscribeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PushKeys": {
        "properties": {
          "p256dh": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "P256Dh",
            "description": "구독 공개키"
          },
          "auth": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Auth",
            "description": "인증 시크릿"
          }
        },
        "type": "object",
        "required": [
          "p256dh",
          "auth"
        ],
        "title": "PushKeys",
        "description": "브라우저가 주는 암호화 키 (RFC 8291)."
      },
...
      "PushSubscriptionIn": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Endpoint",
            "description": "푸시 서비스 엔드포인트 URL"
          },
          "keys": {
            "$ref": "#/components/schemas/PushKeys"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "PushSubscriptionIn",
        "description": "`PushSubscription.toJSON()` 이 그대로 들어온다.\n\n브라우저가 만들어 주는 형태를 그대로 받으므로 클라이언트 코드가 단순해진다."
      },
      "PushSubscriptionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "created": {
            "type": "boolean",
            "title": "Created",
            "description": "새로 만들어졌는지(False 면 이미 있던 구독)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "endpoint",
          "created"
        ],
        "title": "PushSubscriptionOut"
      },
      "PushUnsubscribeIn": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Endpoint"
          }
        },
        "type": "object",
        "required": [
          "endpoint"
        ],
        "title": "PushUnsubscribeIn",
        "description": "해지에는 엔드포인트만 있으면 된다.\n\n여기서는 **호스트를 검사하지 않는다.** 조회만 하고 그 주소로 요청을 보내지\n않으므로 위험이 없고, 목록이 바뀌기 전에 등록된 구독도 해지할 수 있어야 한다."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}